# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Ghar Room Booking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  approved: "approved",
  paid: "paid",
  expired: "expired",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  tenantName: string;
  tenantPhone: string;
  propertyName: string;
  /** @nullable */
  roomNumber: string | null;
  actualRent: number;
  discountedRent: number;
  deposit: number;
  maintenanceFee: number;
  tokenAmount: number;
  stayDurationMonths: number;
  noticePeriodMonths: number;
  status: BookingStatus;
  /** @nullable */
  approvedAt: string | null;
  /** @nullable */
  offerExpiresAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookingBody {
  tenantName: string;
  tenantPhone: string;
  propertyName: string;
  /** @nullable */
  roomNumber?: string | null;
  actualRent: number;
  discountedRent: number;
  deposit: number;
  maintenanceFee: number;
  tokenAmount: number;
  stayDurationMonths: number;
  noticePeriodMonths: number;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  approved: "approved",
  paid: "paid",
  expired: "expired",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingBody {
  tenantName?: string;
  tenantPhone?: string;
  propertyName?: string;
  /** @nullable */
  roomNumber?: string | null;
  actualRent?: number;
  discountedRent?: number;
  deposit?: number;
  maintenanceFee?: number;
  tokenAmount?: number;
  stayDurationMonths?: number;
  noticePeriodMonths?: number;
  status?: UpdateBookingBodyStatus;
}

export interface WhatsappMessage {
  message: string;
  url: string;
  phone: string;
}

export interface BookingStats {
  total: number;
  pending: number;
  approved: number;
  paid: number;
  expired: number;
  cancelled: number;
  totalRevenue: number;
}