# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Ghar Room Booking
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all bookings
 */
export const ListBookingsResponseItem = zod.object({
  id: zod.number(),
  tenantName: zod.string(),
  tenantPhone: zod.string(),
  propertyName: zod.string(),
  roomNumber: zod.string().nullable(),
  actualRent: zod.number(),
  discountedRent: zod.number(),
  deposit: zod.number(),
  maintenanceFee: zod.number(),
  tokenAmount: zod.number(),
  stayDurationMonths: zod.number(),
  noticePeriodMonths: zod.number(),
  status: zod.enum(["pending", "approved", "paid", "expired", "cancelled"]),
  approvedAt: zod.string().nullable(),
  offerExpiresAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListBookingsResponse = zod.array(ListBookingsResponseItem);

/**
 * @summary Create a new booking quotation
 */
export const CreateBookingBody = zod.object({
  tenantName: zod.string(),
  tenantPhone: zod.string(),
  propertyName: zod.string(),
  roomNumber: zod.string().nullish(),
  actualRent: zod.number(),
  discountedRent: zod.number(),
  deposit: zod.number(),
  maintenanceFee: zod.number(),
  tokenAmount: zod.number(),
  stayDurationMonths: zod.number(),
  noticePeriodMonths: zod.number(),
});

/**
 * @summary Get a booking by ID
 */
export const GetBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookingResponse = zod.object({
  id: zod.number(),
  tenantName: zod.string(),
  tenantPhone: zod.string(),
  propertyName: zod.string(),
  roomNumber: zod.string().nullable(),
  actualRent: zod.number(),
  discountedRent: zod.number(),
  deposit: zod.number(),
  maintenanceFee: zod.number(),
  tokenAmount: zod.number(),
  stayDurationMonths: zod.number(),
  noticePeriodMonths: zod.number(),
  status: zod.enum(["pending", "approved", "paid", "expired", "cancelled"]),
  approvedAt: zod.string().nullable(),
  offerExpiresAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a booking
 */
export const UpdateBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookingBody = zod.object({
  tenantName: zod.string().optional(),
  tenantPhone: zod.string().optional(),
  propertyName: zod.string().optional(),
  roomNumber: zod.string().nullish(),
  actualRent: zod.number().optional(),
  discountedRent: zod.number().optional(),
  deposit: zod.number().optional(),
  maintenanceFee: zod.number().optional(),
  tokenAmount: zod.number().optional(),
  stayDurationMonths: zod.number().optional(),
  noticePeriodMonths: zod.number().optional(),
  status: zod
    .enum(["pending", "approved", "paid", "expired", "cancelled"])
    .optional(),
});

export const UpdateBookingResponse = zod.object({
  id: zod.number(),
  tenantName: zod.string(),
  tenantPhone: zod.string(),
  propertyName: zod.string(),
  roomNumber: zod.string().nullable(),
  actualRent: zod.number(),
  discountedRent: zod.number(),
  deposit: zod.number(),
  maintenanceFee: zod.number(),
  tokenAmount: zod.number(),
  stayDurationMonths: zod.number(),
  noticePeriodMonths: zod.number(),
  status: zod.enum(["pending", "approved", "paid", "expired", "cancelled"]),
  approvedAt: zod.string().nullable(),
  offerExpiresAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a booking
 */
export const DeleteBookingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Approve a booking and start the 15-minute timer
 */
export const ApproveBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const ApproveBookingResponse = zod.object({
  id: zod.number(),
  tenantName: zod.string(),
  tenantPhone: zod.string(),
  propertyName: zod.string(),
  roomNumber: zod.string().nullable(),
  actualRent: zod.number(),
  discountedRent: zod.number(),
  deposit: zod.number(),
  maintenanceFee: zod.number(),
  tokenAmount: zod.number(),
  stayDurationMonths: zod.number(),
  noticePeriodMonths: zod.number(),
  status: zod.enum(["pending", "approved", "paid", "expired", "cancelled"]),
  approvedAt: zod.string().nullable(),
  offerExpiresAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get a pre-filled WhatsApp message for a booking
 */
export const GetWhatsappMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWhatsappMessageResponse = zod.object({
  message: zod.string(),
  url: zod.string(),
  phone: zod.string(),
});

/**
 * @summary Get booking summary stats
 */
export const GetBookingStatsResponse = zod.object({
  total: zod.number(),
  pending: zod.number(),
  approved: zod.number(),
  paid: zod.number(),
  expired: zod.number(),
  cancelled: zod.number(),
  totalRevenue: zod.number(),
});