# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Ghar Room Booking
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  tenantName: string;
  tenantPhone: string;
  propertyName: string;
  /** @nullable */
  roomNumber: string | null;
  actualRent: number;
  discountedRent: number;
  deposit: number;
  maintenanceFee: number;
  tokenAmount: number;
  stayDurationMonths: number;
  noticePeriodMonths: number;
  status: BookingStatus;
  /** @nullable */
  approvedAt: string | null;
  /** @nullable */
  offerExpiresAt: string | null;
  createdAt: string;
  updatedAt: string;
}