# lib/api-zod/src/generated/types/updateBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Ghar Room Booking
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookingBodyStatus } from "./updateBookingBodyStatus";

export interface UpdateBookingBody {
  tenantName?: string;
  tenantPhone?: string;
  propertyName?: string;
  /** @nullable */
  roomNumber?: string | null;
  actualRent?: number;
  discountedRent?: number;
  deposit?: number;
  maintenanceFee?: number;
  tokenAmount?: number;
  stayDurationMonths?: number;
  noticePeriodMonths?: number;
  status?: UpdateBookingBodyStatus;
}